* toplev.c (do_compile):  Remove unwanted push_srcloc/pop_srcloc.
	* opts.c (common_handle_option):  Remove bogus cast.


git-svn-id: svn+ssh://gcc.gnu.org/svn/gcc/branches/compile-server-branch@71340 138bc75d-0d04-0410-961f-82ee72b054a4

+++ gcc/opts.c
@@ -965,7 +965,7 @@ common_handle_option (size_t scode, const char *arg,
 	  else
 	    {
 	      int i = integral_argument (arg);
-	      if (i == (unsigned int) -1 || i > 2)
+	      if (i == -1 || i > 2)
 		return 0;
 	      server_mode = i;
 	    }
